requirements.txt
keyboard
Pillow
pystray
pyinstaller
requests
packaging